First Commit,  Need to Clean up Class and Add Demos
This is the very first commit,  I created a simple Extension Class for
making Photo Gallery's in my last web development project that required
the use of many Collections, Photostreams, etc so I wrote this class.  I
will be updating it and adding the CSS and a Demo Page for it in the
very near future.